requirements.txt
# Web框架和API
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.5.0
python-multipart==0.0.6

# 数据库和缓存
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1

# 搜索引擎
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0

# 向量数据库
qdrant-client==1.7.0

# 爬虫框架
scrapy==2.11.0
scrapy-redis==0.7.3
scrapy-splash==0.8.0

# 任务队列
celery==5.3.4
celery[redis]==5.3.4
flower==2.0.1

# 机器学习和NLP
torch==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
jieba==0.42.1

# 数据处理
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 工具库
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
loguru==0.7.2
python-dateutil==2.8.2

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
isort==5.12.0

# 监控和日志
prometheus-client==0.19.0
structlog==23.2.0

# 安全
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
